requirements.txt
# Core Dependencies
Flask
Flask-Login
Flask-WTF
Flask-Principal
WTForms
Werkzeug
python-dotenv
Jinja2

# Parsing and Data Processing
spacy
pandas
mail-parser
validators
pyyaml  # For YAML configuration loading

# JSON and Schema Validation
jsonschema
json-log-formatter

# HTTP Requests
requests

# Logging and Utilities
tenacity
filelock

# Environment and Configuration
python-multipart

# Security
bcrypt
passlib[bcrypt]
itsdangerous
python-jose[cryptography]

# Database
SQLAlchemy
psycopg2-binary

# Scheduling
APScheduler

# Natural Language Processing
nltk
torch
transformers
langchain
openai
spacy-transformers  # For improved spaCy integration with transformers

# Fuzzy Matching
fuzzywuzzy
python-Levenshtein  # Optional, but improves fuzzywuzzy performance

# Development and Testing
pytest
pytest-mock
unittest2
flake8
black
isort
flake8-black
sphinx

# Progress Bars
tqdm

# Numerical Computing
numpy

# Note:
# spaCy models like 'en-core-web-sm' need to be downloaded separately.
# After installing dependencies, run the following command to download the model:
# python -m spacy download en_core_web_sm
